Heartbeats should only reschedule existing heartbeat jobs
