+++ apps/webapp/app/v3/marqs/v3VisibilityTimeout.server.ts
@@ -3,6 +3,10 @@ import { TaskRunHeartbeatFailedService } from "../taskRunHeartbeatFailed.server"
 import { VisibilityTimeoutStrategy } from "./types";
 
 export class V3GraphileVisibilityTimeout implements VisibilityTimeoutStrategy {
+  async startHeartbeat(messageId: string, timeoutInMs: number): Promise<void> {
+    await TaskRunHeartbeatFailedService.enqueue(messageId, new Date(Date.now() + timeoutInMs));
+  }
+
   async heartbeat(messageId: string, timeoutInMs: number): Promise<void> {
     await TaskRunHeartbeatFailedService.enqueue(messageId, new Date(Date.now() + timeoutInMs));
   }
@@ -13,13 +17,20 @@ export class V3GraphileVisibilityTimeout implements VisibilityTimeoutStrategy {
 }
 
 export class V3LegacyRunEngineWorkerVisibilityTimeout implements VisibilityTimeoutStrategy {
-  async heartbeat(messageId: string, timeoutInMs: number): Promise<void> {
+  async startHeartbeat(messageId: string, timeoutInMs: number): Promise<void> {
     await legacyRunEngineWorker.enqueue({
       id: `heartbeat:${messageId}`,
       job: "runHeartbeat",
       payload: { runId: messageId },
       availableAt: new Date(Date.now() + timeoutInMs),
     });
+  }
+
+  async heartbeat(messageId: string, timeoutInMs: number): Promise<void> {
+    await legacyRunEngineWorker.reschedule(
+      `heartbeat:${messageId}`,
+      new Date(Date.now() + timeoutInMs)
+    );
   }
 
   async cancelHeartbeat(messageId: string): Promise<void> {
